requirements.txt
# Core API Dependencies
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
pydantic==2.11.7

# Core LLM and Vector Database Dependencies
langchain==0.1.5
langchain-google-genai==0.0.11
langchain-pinecone==0.0.3
pinecone-client==3.0.3
google-generativeai>=0.4.1

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.0
python-docx==1.1.0

# Data Processing and ML
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0

# Environment and Configuration
python-dotenv==1.0.0
typing-extensions==4.9.0

# API and Web Framework
requests==2.31.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4

# Logging and Monitoring
loguru==0.7.2

# JSON and Data Validation
jsonschema==4.20.0

# Progress bars and utilities
tqdm==4.66.1
